runtime: Move the resourcecontrol package one layer up
And try to reduce the number of virtcontainers packages, step by step.
